PYTHON/actividad_U4/requirements.txt
Django
requests
djangorestframework
mysqlclient